Atemmpt tidy fix

+++ src/lib/utils/date_time_utils.cpp
@@ -32,13 +32,13 @@ boost::gregorian::date date_interval(const boost::gregorian::date& start_date, i
     case DatetimeComponent::Year: {
       // We obtain an iterator that adds offset years when incremented. Thus, we have to actually increment and
       // dereference it. The same applies to the other cases.
-      return *(++boost::gregorian::year_iterator(start_date, offset));
+      return *(++boost::gregorian::year_iterator(start_date, static_cast<int>(offset)));
     }
     case DatetimeComponent::Month: {
-      return *(++boost::gregorian::month_iterator(start_date, offset));
+      return *(++boost::gregorian::month_iterator(start_date, static_cast<int>(offset)));
     }
     case DatetimeComponent::Day: {
-      return *(++boost::gregorian::day_iterator(start_date, offset));
+      return *(++boost::gregorian::day_iterator(start_date, static_cast<int>(offset)));
     }
     default:
       Fail("Invalid time unit for date interval: " + std::string{magic_enum::enum_name(unit)});
